Connecting the Frontend to AdminChannel
We listen to the presence_state and presence_diff events
so we can keep track of and sync presences in the frontend.
We then store these presences in our Chat component's
state. That way, our component will rerender when
presences' state changes. We pass these presences to
Sidebar, which is a stateless functional component, and
only deals with rendering of the presences.