Admin receives chat messages and changes chat rooms.
An admin can now click on a user on the sidebar and
view messages from that user. A user can send messages
and the admin will receive them in real-time. Since
our messages are already persisted in the DB, admin
will be able to see them every time they join a chatroom.

Next, we will need admin to be able to respond to users.