Admin can respond in chat.
We handle input entered by admin and send the input to
the backend on when they press 'Enter'. Since we already
have the logic in place for sending this message to all
members of a room, the user will receive this message
once it has been saved and broadcast by our server.