Replace hardcoded values with config in webpack.
The host for our JSON API and Websocket servers will
be different in our production environment. This needs
to be configurable. In Webpack, we can define values
during build time which can be configured to be different
depending on your environment. Later on, we will add
a webpack config specifically for production so we can
easily create a production build and deploy it.