List active and inactive users.
Using presences alone to populate admin's sidebar
doesn't allow us to see inactive users. To deal with
this, we populate our list with users that we store in
our server. Since our server passes a list of anonymous
users when we connect to admin:active_users, we can use
this list to fill up our sidebar with an initial list
of users.

To handle updating this list, we listen to the lobby_list
event. Through this event, the server will pass us users
that have connected to our app. Once connected, we check
if that user exists in our list and add them if they
are new.